🔨refactor: rename task initialization, move custom model to core, updated exceptions
