+++ tasks/__init__.py
@@ -1,4 +1,4 @@
-def initialize_initialization(task, subtask=None):
+def export_task_modules(task, subtask=None):
     if task == "text_generation":
         from nyuntam_adapt.tasks.text_generation_hf import CausalLLM
