+++ tasks/question_answering_hf/task.py
@@ -10,7 +10,7 @@ from transformers import (
     default_data_collator,
     BitsAndBytesConfig,
 )
-from nyuntam_adapt.tasks.custom_model import prepare_custom_model_support
+from nyuntam_adapt.core.custom_model import prepare_custom_model_support
 from nyuntam_adapt.core.base_task import BaseTask
 from nyuntam_adapt.utils.task_utils import prepare_model_for_kbit_training, ModelLoadingError
 
@@ -216,7 +216,9 @@ class QuestionAnswering(BaseTask):
                         **self.model_args,
                     )
             except Exception as e:
-                raise ModelLoadingError(f"Following Error Happened : {e}") from e
+                raise ModelLoadingError(
+                    f"Model ({self.model_path}) cannot be loaded due to : {e}, \n Maybe wrong name?"
+                ) from e
 
         if use_bnb:
             model = prepare_model_for_kbit_training(model, self.gradient_checkpointing)
