+++ tasks/text_classification_hf/task.py
@@ -9,7 +9,7 @@ from transformers import (
     AutoModelForTokenClassification,
     DataCollatorForTokenClassification,
 )
-from nyuntam_adapt.tasks.custom_model import prepare_custom_model_support
+from nyuntam_adapt.core.custom_model import prepare_custom_model_support
 from nyuntam_adapt.core.base_task import BaseTask
 from nyuntam_adapt.utils.task_utils import ModelLoadingError
 
@@ -95,7 +95,9 @@ class SequenceClassification(BaseTask):
                             use_flash_attention_2=self.flash_attention,
                         )
                 except Exception as e:
-                    raise ModelLoadingError(f"Following Error Happened : {e}") from e
+                    raise ModelLoadingError(
+                        f"Model ({self.model_path}) cannot be loaded due to : {e}, \n Maybe wrong name?"
+                    ) from e
 
             self.collate_fn = DataCollatorForTokenClassification(
                 tokenizer=self.tokenizer
@@ -160,7 +162,9 @@ class SequenceClassification(BaseTask):
                         )
 
                 except Exception as e:
-                    raise ModelLoadingError(f"Following Error Happened : {e}") from e
+                    raise ModelLoadingError(
+                        f"Model ({self.model_path}) cannot be loaded due to : {e}, \n Maybe wrong name?"
+                    ) from e
 
             self.collate_fn = DataCollatorWithPadding(tokenizer=self.tokenizer)
             self.compute_metrics = self.compute_metrics_sequence
