+++ tasks/translation_hf/task.py
@@ -98,7 +98,9 @@ class Translation(BaseTask):
                         use_flash_attention_2=self.flash_attention,
                     )
             except Exception as e:
-                raise ModelLoadingError(f"Following Error Happened : {e}") from e
+                raise ModelLoadingError(
+                    f"Model ({self.model_path}) cannot be loaded due to : {e}, \n Maybe wrong name?"
+                ) from e
 
         if use_bnb:
             model = prepare_model_for_kbit_training(model, self.gradient_checkpointing)
